plugin/reload move test case to dedicated package
